Respect paths from IAppDataService
