+++ src/Orchestra.Shell.Shared/Services/ApplicationInitializationServiceBase.cs
@@ -10,6 +10,7 @@ namespace Orchestra.Services
     using System;
     using System.Threading.Tasks;
     using Catel.IoC;
+    using Catel.Logging;
     using Catel.Threading;
     using MethodTimer;
 
@@ -21,6 +22,8 @@ namespace Orchestra.Services
 
         public virtual async Task InitializeBeforeShowingSplashScreenAsync()
         {
+            InitializeLogging();
+
             var serviceLocator = this.GetServiceLocator();
             var themeService = serviceLocator.ResolveType<IThemeService>();
 
@@ -53,6 +56,20 @@ namespace Orchestra.Services
         protected static async Task RunAndWaitAsync(params Func<Task>[] actions)
         {
             await TaskHelper.RunAndWaitAsync(actions);
+        }
+
+        protected virtual void InitializeLogging()
+        {
+            LogHelper.CleanUpAllLogTypeFiles();
+
+            var fileLogListener = LogHelper.CreateFileLogListener(LogFilePrefixes.EntryAssemblyName);
+
+            fileLogListener.IsDebugEnabled = false;
+            fileLogListener.IsInfoEnabled = true;
+            fileLogListener.IsWarningEnabled = true;
+            fileLogListener.IsErrorEnabled = true;
+
+            LogManager.AddListener(fileLogListener);
         }
     }
 }
